dom/tests/mochitest/chrome/DOMWindowCreated_chrome.xhtml

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css" type="text/css"?>
<!--
  Test for Persistent Storage in chrome
  -->
<window id="sample-window" width="400" height="400"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script type="application/javascript" 
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" 
          src="chrome://mochikit/content/chrome-harness.js"></script>

<script type="application/javascript">
var SimpleTest = window.arguments[0].SimpleTest;

document.addEventListener("DOMWindowCreated", function(e) {
  var otherWindow = e.target.defaultView.wrappedJSObject;
  SimpleTest.is(e.type, "DOMWindowCreated", "DOMWindowCreated: " + otherWindow);
  otherWindow.doneFunction = function() {
    SimpleTest.ok(true, "doneFunction was called");
    SimpleTest.finish();
    window.close();
  };
});

var root = getRootDirectory(window.location.href);
var el = document.createXULElement("iframe");
el.setAttribute('type', 'content');
el.setAttribute('src', root + 'DOMWindowCreated_content.html');
document.documentElement.appendChild(el);
</script>
</window>
